backend/requirements.txt
torch>=1.9.0
torchvision>=0.15.2
torchaudio>=2.0.2
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.23.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-dotenv>=0.19.0
scikit-learn>=1.3.0
pandas>=2.1.1
numpy>=1.21.0
pyttsx3>=2.90
requests>=2.31.0
wandb>=0.15.12
tqdm>=4.66.1
seaborn>=0.13.0
matplotlib>=3.8.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
aiofiles>=0.7.0 
